Reorganise manage public holidays
