+++ src/components/content/hr/HRSpecific/publicHolidays/publicHolidays.css
@@ -6,10 +6,18 @@
   visibility: hidden !important;
 }
 
-.PublicFormStyle {
-  width: 60%;
+/* .PublicFormStyle {
+  width: 100%;
   margin: 3em auto 0em auto;
   border: 0.1em solid #003366;
   padding: 5em 5em 1em 5em;
   border-radius: 0.3em;
+} */
+
+.inlineItem{
+  display: inline;
+}
+
+.positionBtn{
+  margin-left: 20%;
 }
